requirements.txt
pandas==2.1.3
psycopg2-binary==2.9.9
numpy==1.26.2
python-dotenv==1.0.0
tabulate==0.9.0
matplotlib==3.8.2
scipy==1.11.4

